fix and test

+++ models/official/resnet/onecycle.py
@@ -56,4 +56,4 @@ class OneCycleScheduler():
 def lrs(step,total_step):
     low = math.log2(1e-5)
     high = math.log2(10)
-    return tf.Tensor(2**(low+(high-low)*step/total_step))
+    return tf.Tensor(2**(low+(high-low)*step/total_step),dtype=tf.float32)
